+++ src/3rdparty/squirrel/squirrel/sqmem.cpp
@@ -2,8 +2,8 @@
 	see copyright notice in squirrel.h
 */
 #include "sqpcheader.h"
-void *sq_vm_malloc(SQUnsignedInteger size){	return malloc(size); }
+void *sq_vm_malloc(SQUnsignedInteger size){	return malloc((size_t)size); }
 
-void *sq_vm_realloc(void *p, SQUnsignedInteger oldsize, SQUnsignedInteger size){ return realloc(p, size); }
+void *sq_vm_realloc(void *p, SQUnsignedInteger oldsize, SQUnsignedInteger size){ return realloc(p, (size_t)size); }
 
 void sq_vm_free(void *p, SQUnsignedInteger size){	free(p); }
